+++ crates/config/src/schema/generator.rs
@@ -5,6 +5,8 @@ use std::collections::HashSet;
 use std::fs;
 use std::path::Path;
 
+/// A generator collects [`SchemaType`]s and renders them to a specific file,
+/// using a renderer that implements [`SchemaRenderer`].
 #[derive(Default)]
 pub struct SchemaGenerator {
     references: HashSet<String>,
@@ -12,10 +14,13 @@ pub struct SchemaGenerator {
 }
 
 impl SchemaGenerator {
+    /// Add a [`SchemaType`] to be rendered, derived from the provided [`Schematic`].
     pub fn add<T: Schematic>(&mut self) {
         self.add_schema(T::generate_schema());
     }
 
+    /// Add an explicit [`SchemaType`] to be rendered, and recursively add any nested schemas.
+    /// Schemas with a name will be considered a reference.
     pub fn add_schema(&mut self, schema: SchemaType) {
         // Store the name so that we can use it as a reference for other types
         if let Some(name) = schema.get_name() {
@@ -24,7 +29,7 @@ impl SchemaGenerator {
                 return;
             }
 
-            self.references.insert(name.clone());
+            self.references.insert(name.to_owned());
 
         // Types without a name cannot be rendered at the root
         } else {
@@ -65,6 +70,8 @@ impl SchemaGenerator {
         self.schemas.push(schema);
     }
 
+    /// Generate an output by rendering all collected [`SchemaType`]s using the provided
+    /// [`SchemaRenderer`], and finally write to the provided file path.
     pub fn generate<P: AsRef<Path>, R: SchemaRenderer>(
         &self,
         output_file: P,