+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 
 #### 🚀 Updates
 
-- Added a new schema layer that defines the structure of built-in Rust types, and schematic
+- Added a new schema layer that defines the structure of built-in Rust types and schematic
   configuration types.
   - Implements the new `Schematic` trait.
   - Types provided by the new `schematic_types` crate.